+++ R/RcppExports.R
@@ -61,6 +61,10 @@ node_children <- function(node, onlyNode = TRUE) {
     .Call('xml2_node_children', PACKAGE = 'xml2', node, onlyNode)
 }
 
+node_length <- function(node, onlyNode = TRUE) {
+    .Call('xml2_node_length', PACKAGE = 'xml2', node, onlyNode)
+}
+
 node_parents <- function(node) {
     .Call('xml2_node_parents', PACKAGE = 'xml2', node)
 }