+++ NEWS.md
@@ -1,2 +1,3 @@
 # xml2 0.1.0.9000
 
+* Add `xml_length()` to count the number of children (#32).